% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggbeeswarm.R
\name{digits2number}
\alias{digits2number}
\title{Convert a vector of integers representing digits in an arbitrary base to an integer}
\usage{
digits2number(digits, base, fractional = FALSE)
}
\arguments{
\item{digits}{a vector of integers representing digits in an arbitrary base}

\item{base}{the base for the numeral system (e.g. 2 for binary or 8 for octal)}

\item{fractional}{divide the}
}
\value{
an integer
}
\description{
Takes a vector of integers representing digits in an arbitrary base e.g. binary or octal and converts it into an integer (or the integer divided by base^length(digits) for the number of digits if fractional is TRUE). Note that the first digit in the input is the least significant.
}
\examples{
digits2number(c(4,4,1),8)
digits2number(number2digits(100))
}
\references{
\url{https://en.wikipedia.org/wiki/Radix}
}

