% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-beeswarm.R
\name{geom_beeswarm}
\alias{geom_beeswarm}
\title{Points, jittered to reduce overplotting using the beeswarm package}
\usage{
geom_beeswarm(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  ...,
  method = "swarm",
  cex = 1,
  side = 0L,
  priority = "ascending",
  fast = TRUE,
  dodge.width = NULL,
  corral = "none",
  corral.width = 0.9,
  groupOnX = NULL,
  beeswarmArgs = list(),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{method}{Method for arranging points (see Details below)}

\item{cex}{Scaling for adjusting point spacing (see \code{\link[beeswarm]{swarmx}}).
Values between 1 (default) and 3 tend to work best.}

\item{side}{Direction to perform jittering: 0: both directions; 
1: to the right or upwards; -1: to the left or downwards.}

\item{priority}{Method used to perform point layout (see Details below)}

\item{fast}{Use compiled version of swarm algorithm? This option is ignored 
for all methods expect `"swarm"` and `"compactswarm"`.}

\item{dodge.width}{Amount by which points from different aesthetic groups 
will be dodged. This requires that one of the aesthetics is a factor.}

\item{corral}{`string`. Method used to adjust points that would be placed to
wide horizontally, default is `"none"`. See details below.}

\item{corral.width}{`numeric`. Width of the corral, default is `0.9`.}

\item{groupOnX}{Deprecated.}

\item{beeswarmArgs}{Deprecated.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
The beeswarm geom is a convenient means to offset points within categories to
 reduce overplotting. Uses the beeswarm package
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "point")}
}

\examples{

  ggplot2::qplot(class, hwy, data = ggplot2::mpg, geom='beeswarm')
  # Generate fake data
  distro <- data.frame(
    'variable'=rep(c('runif','rnorm'),each=100),
    'value'=c(runif(100, min=-3, max=3), rnorm(100))
  )
  ggplot2::qplot(variable, value, data = distro, geom='beeswarm')
  ggplot2::ggplot(distro,aes(variable, value)) +
    geom_beeswarm(priority='density',size=2.5)
}
\seealso{
\code{\link{geom_quasirandom}} an alternative method,
 \code{\link[beeswarm]{swarmx}} how spacing is determined,
 \code{\link[ggplot2]{geom_point}} for regular, unjittered points,
 \code{\link[ggplot2]{geom_jitter}} for jittered points,
 \code{\link[ggplot2]{geom_boxplot}} for another way of looking at the conditional
    distribution of a variable
}
