% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-beeswarm.R
\docType{data}
\name{GeomBeeswarm}
\alias{GeomBeeswarm}
\title{geom object for use in geom_beeswarm}
\format{An object of class \code{GeomBeeswarm} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
GeomBeeswarm
}
\description{
geom object for use in geom_beeswarm
}
\keyword{datasets}
