% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-beeswarm.R
\name{makeContent.beeswarm}
\alias{makeContent.beeswarm}
\title{grid::makeContent function for the grobTree of GeomBeeswarm objects}
\usage{
makeContent.beeswarm(self)
}
\arguments{
\item{self}{a beeswarm gridTree grob}
}
\value{
a pointsGrob
}
\description{
grid::makeContent function for the grobTree of GeomBeeswarm objects
}
