% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggbeeswarm.R
\name{offset_x}
\alias{offset_x}
\title{Offset data to avoid overplotting.}
\usage{
offset_x(y, x, width = 0.4, var_width = FALSE, adjust = 0.5,
  nbins = 1000)
}
\arguments{
\item{y}{vector of data points}

\item{x}{a grouping factor for y (i.e. what will be the x axis)}

\item{width}{the maximum x-width of each group (0 = no width, 1 = max; default = 0.4)}

\item{var_width}{adjust the width of each group based on the number of points in the group}

\item{adjust}{bandwidth used to adjust the density}

\item{nbins}{the number of points used to calculate density}
}
\description{
Offset data to avoid overplotting.
}

