% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-quasirandom.R
\name{position_quasirandom}
\alias{position_quasirandom}
\title{Violin point-style plots to show overlapping points. x must be discrete.}
\usage{
position_quasirandom(
  method = "quasirandom",
  width = NULL,
  varwidth = FALSE,
  bandwidth = 0.5,
  nbins = NULL,
  dodge.width = 0,
  groupOnX = NULL
)
}
\arguments{
\item{method}{the method used for distributing points
(quasirandom, pseudorandom, smiley or frowney)}

\item{width}{the maximum amount of spread (default: 0.4)}

\item{varwidth}{vary the width by the relative size of each group}

\item{bandwidth}{the bandwidth adjustment to use when calculating density
Smaller numbers (< 1) produce a tighter "fit". (default: 0.5)}

\item{nbins}{the number of bins used when calculating density
(has little effect with quasirandom/random distribution)}

\item{dodge.width}{Amount by which points from different aesthetic groups 
will be dodged. This requires that one of the aesthetics is a factor.}

\item{groupOnX}{Deprecated.}
}
\description{
Violin point-style plots to show overlapping points. x must be discrete.
}
\seealso{
\code{\link[vipor]{offsetSingleGroup}}, \code{\link{geom_quasirandom}}

Other position adjustments: 
\code{\link{offset_beeswarm}()}
}
\concept{position adjustments}
