% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/position-violinPoint.R
\name{position_violinPoint}
\alias{position_violinPoint}
\title{Violin point-style plots to show overlapping points. x must be discrete.}
\usage{
position_violinPoint(width = NULL, varwidth = NULL, bandwidth = NULL)
}
\arguments{
\item{width}{the maximum amount of spread (default: 0.4)}

\item{varwidth}{vary the width by the relative size of each group}

\item{bandwidth}{the bandwidth adjustment to use when calculating density
Smaller numbers (< 1) produce a tighter "fit". (default: 0.5)}
}
\description{
Violin point-style plots to show overlapping points. x must be discrete.
}
\examples{
if(requireNamespace("ggplot2",quietly=TRUE)&requireNamespace("proto",quietly=TRUE)){
#  ggplot2::qplot(class, hwy, data = ggplot2::mpg, position=position_violinPoint())
  # Generate fake data
  distro <- data.frame(
    'variable'=rep(c('runif','rnorm'),each=100),
    'value'=c(runif(100, min=-3, max=3), rnorm(100))
  )
#  ggplot2::qplot(variable, value, data = distro, position = position_violinPoint())
#  ggplot2::qplot(variable, value, data = distro, position = position_violinPoint(width=0.1))
}
}

