% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggbeeswarm.R
\name{vanDerCorput}
\alias{vanDerCorput}
\title{Generate van der Corput sequences}
\usage{
vanDerCorput(n, base = 2)
}
\arguments{
\item{n}{the first n digits of the van der Corput sequence}

\item{base}{the base to use for calculating the van der Corput sequence}
}
\description{
Generate van der Corput sequences
}

