% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-beeswarm.R
\name{offset_beeswarm}
\alias{offset_beeswarm}
\title{An internal function to calculate new positions for geom_beeswarm}
\usage{
offset_beeswarm(
  data,
  xRange = 1,
  yRange = 1,
  priority = c("ascending", "descending", "density", "random", "none"),
  cex = 1,
  groupOnX = NULL,
  dodge.width = 0,
  beeswarmArgs = list(),
  oSize = c(1/200, 1/200)
)
}
\arguments{
\item{data}{A data.frame containing plotting data in columns x and y. Usually obtained from data processed by ggplot2.}

\item{xRange}{span of the x axis e.g. yMax-yMin}

\item{yRange}{span of the y axis e.g. yMax-yMin}

\item{priority}{Method used to perform point layout (see \code{\link{swarmx}})}

\item{cex}{Scaling for adjusting point spacing (see \code{\link{swarmx}})}

\item{groupOnX}{if TRUE then jitter is added to the x axis and if FALSE jitter is added to the y axis. Prior to v0.6.0, the default NULL causes the function to guess which axis is the categorical one based on the number of unique entries in each. This could result in unexpected results when the x variable has few unique values and so in v0.6.0 the default was changed to always jitter on the x axis unless groupOnX=FALSE. Also consider \code{\link[ggplot2]{coord_flip}}.}

\item{dodge.width}{Amount by which points from different aesthetic groups will be dodged. This requires that one of the aesthetics is a factor.}

\item{beeswarmArgs}{A list of additional arguments to be passed to the \link{swarmx} function of beeswarm e.g. \code{list(side=1)} or \code{list(side=-1)} to only distribute points to the right/left side}

\item{oSize}{A 2 element vector giving the width and height of a cex=1 character "o" in user coordinates}
}
\description{
An internal function to calculate new positions for geom_beeswarm
}
\seealso{
\code{\link{geom_beeswarm}}, \code{\link{position_quasirandom}}, \code{\link[beeswarm]{swarmx}}

Other position adjustments: 
\code{\link{position_quasirandom}()}
}
\concept{position adjustments}
