% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/position-beeswarm.r
\name{position-beeswarm}
\alias{position-beeswarm}
\alias{position_beeswarm}
\title{Beeswarm-style plots to show overlapping points. x must be discrete.}
\usage{
position_beeswarm(width = NULL, nbins = NULL)
}
\arguments{
\item{width}{the maximum amount of spread (default: 0.4)}

\item{nbins}{number of points to use to estimate point density (default: 128)}
}
\description{
Beeswarm-style plots to show overlapping points. x must be discrete.
}
\examples{
qplot(class, hwy, data = mpg, position="beeswarm")
# Generate fake data
distro <- melt(data.frame(list(runif=runif(100, min=-3, max=3), rnorm=rnorm(100))))
qplot(variable, value, data = distro, position = "beeswarm")
# Spacing and number of bins can be adjusted
qplot(variable, value, data = distro, position = position_beeswarm(width=0.05, nbins=35))
}

