% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/position-beeswarm.r
\name{position_beeswarm}
\alias{position_beeswarm}
\title{Beeswarm-style plots to show overlapping points. x must be discrete.}
\usage{
position_beeswarm(width = NULL, var_width = NULL, bandwidth = NULL)
}
\arguments{
\item{width}{the maximum amount of spread (default: 0.4)}

\item{var_width}{vary the width by the size of the data}

\item{bandwidth}{the bandwidth adjustment to use when calculating density
Smaller numbers (< 1) produce a tighter "fit". (default: 0.5)}
}
\description{
Beeswarm-style plots to show overlapping points. x must be discrete.
}
\examples{
qplot(class, hwy, data = mpg, position=position_beeswarm())
# Generate fake data
distro <- melt(data.frame(list(runif=runif(100, min=-3, max=3), rnorm=rnorm(100))))
qplot(variable, value, data = distro, position = position_beeswarm())
qplot(variable, value, data = distro, position = position_beeswarm(width=0.1))
}

